fix `make manifest` build failure
Move quote parsing logic into a separate package which does not include
any additional headers. This package could be used by the controller
package.
